save userId that suggested org
