+++ packages/api/router/organization/mutation.createNewSuggestion.handler.ts
@@ -1,4 +1,4 @@
-import { getAuditedClient } from '@weareinreach/db'
+import { generateId, getAuditedClient } from '@weareinreach/db'
 import { type TRPCHandlerParams } from '~api/types/handler'
 
 import { type TCreateNewSuggestionSchema } from './mutation.createNewSuggestion.schema'
@@ -8,7 +8,32 @@ const createNewSuggestion = async ({
 	input,
 }: TRPCHandlerParams<TCreateNewSuggestionSchema, 'protected'>) => {
 	const prisma = getAuditedClient(ctx.actorId)
-	const result = await prisma.suggestion.create(input)
+	const { countryId, orgName, orgSlug, communityFocus, orgAddress, orgWebsite, serviceCategories } = input
+	const organizationId = generateId('organization')
+
+	const result = await prisma.suggestion.create({
+		data: {
+			organization: {
+				create: {
+					id: organizationId,
+					name: orgName,
+					slug: orgSlug,
+					source: { connect: { source: 'suggestion' } },
+				},
+			},
+			data: {
+				orgWebsite,
+				orgAddress,
+				countryId,
+				communityFocus,
+				serviceCategories,
+			},
+			suggestedBy: { connect: { id: ctx.actorId } },
+		},
+		select: {
+			id: true,
+		},
+	})
 	return result
 }
 export default createNewSuggestion
